First pass at documentation overhaul
