+++ .github/workflows/serato.yaml
@@ -8,7 +8,7 @@
 
 name: serato-reader
 
-on: [push, pull_request]  # yamllint disable-line rule:truthy
+on: [push]  # yamllint disable-line rule:truthy
 
 jobs:
   serato-mac:
