+++ .github/workflows/yetus.yaml
@@ -1,7 +1,7 @@
 ---
 name: Static Linting
 
-on: [push, pull_request]  # yamllint disable-line rule:truthy
+on: [push]  # yamllint disable-line rule:truthy
 
 jobs:
   build:
