+++ docs/.gitignore
@@ -1,0 +1,3 @@
+*.swp
+/_build
+/doctrees
