+++ docs/comparisons.rst
@@ -1,0 +1,52 @@
+Versus Comparisons
+==================
+
+There are a lot of different solutions to provide titling for your stream. While
+we would love for you to choose and and work with us, we are realistic that this software
+may not be the one for you.
+
+Serato's Now Playing Twitch Extension
+-------------------------------------
+
+.. list-table::
+   :header-rows: 1
+
+   * - Description
+     - This App
+     - Serato Extension
+   * - Works locally
+     - ✅
+     - 🚫
+   * - Works with Live Playlists
+     - ✅
+     - ✅
+   * - Cover art
+     - ✅
+     - 🚫
+   * - Multi-DJ on one stream friendly
+     - ✅
+     - 🚫
+   * - Album name can be displayed
+     - ✅
+     - 🚫
+   * - Label/Publisher can be displayed
+     - ✅
+     - 🚫
+   * - High degree of customization via templates
+     - ✅
+     - 🚫
+   * - Open source
+     - ✅
+     - 🚫
+   * - Supports multi-deck
+     - ✅
+     - 🚫
+   * - Ability to give data to other programs to manipulate
+     - ✅
+     - 🚫
+   * - Supports 2-computer streaming
+     - ✅
+     - ✅
+   * - Supports Chat Announcements
+     - ✅
+     - 🚫
