+++ docs/index.rst
@@ -1,0 +1,41 @@
+.. Now Playing documentation master file, created by
+   sphinx-quickstart on Sat May 29 21:24:23 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to Now Playing's documentation!
+=======================================
+
+**Now Playing** is a tool written in Python to retrieve the current/last played song in Serato DJ,
+Virtual DJ and other software that writes M3U files, or MPRIS2-compatible software.
+
+Compiled, standalone versions are available for:
+
+* Windows
+* macOS (10.13/High Sierra to 10.15/Catalina)
+
+For everyone else, you will need to build and install locally.  See `Developers <#Developers>`_ below.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   installation
+   usage
+   settings
+   input/m3u
+   input/mpris2
+   input/serato
+   output/obswebsocket
+   output/twitchbot
+   output/webserver
+   templatevariables
+   developers
+   comparisons
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
