+++ docs/input/m3u.rst
@@ -1,0 +1,29 @@
+M3U Support
+===========
+
+M3U is a fairly standardized file format that is used by a variety of software, such as
+Virtual DJ.
+
+Instructions
+------------
+
+#. Open Settings from the Now Playing icon
+#. Select Input Source from left-hand column
+
+.. image:: images/m3u-source-selection.png
+   :target: images/m3u-source-selection.png
+   :alt: M3U Source Selection
+
+#. Select the M3U from the list of available input sources.
+#. Select M3U from the left-hand column.
+#. Enter or, using the button, select the directory where your m3u files will appear.
+   For Virtual DJ, this location should be your Documents/Virtual DJ/History directory.
+
+.. image:: images/m3u-source.png
+   :target: images/m3u-source.png
+   :alt: M3U Directory Selection
+
+#. Click Save
+
+Now Playing will now read that directory, reading the most current m3u file for the
+currently playing track information.
