+++ docs/requirements.txt
@@ -1,0 +1,1 @@
+sphinx==4.0.2
