+++ docs/usage.rst
@@ -1,0 +1,11 @@
+Usage
+=====
+
+Start the Now Playing app from the location where you placed it.The app is controlled and
+configured by accessing the menu from the icon in the Windows system tray or Mac
+menu bar after configuration.
+
+However, the first time you run the app a [Settings](#settings) window will appear.
+Populate the fields accordingly, and press Save. Once saved, the app will start polling
+for new songs. The app can be controlled and exited from the Windows system task tray or
+the macOS menu bar icon.
